agents/python/paper-executor/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "obsidian-paper-executor"
version = "0.1.0"
description = "M0 paper-trading executor against Alpaca."
requires-python = ">=3.12"
dependencies = [
    "obsidian-agent",
    "alpaca-py>=0.30",
]

[tool.setuptools.packages.find]
where = ["src"]
